sandboxedmodule is ready for production.
