+++ index.js
@@ -1,21 +1,17 @@
 /* global WebInspector TimelineModelTreeView */
 'use strict';
 
-//require('mock-globals') // nope. doesn't sandbox Natives
-// var Sandbox = require('sandbox'); // nope. doesn't have a global obj
-// vm.createContext // nope. doesnt have global.
 var fs = require('fs');
 var vm = require('vm')
 
-
-var WebInspector = {}
-class TraceToTimelineModel {
+class ModelAPI {
 
   constructor(events) {
 
+    // Everything happens in a sandboxed vm context, to keep globals and natives separate.
     var glob = { require: require, global: global, console: console, process, process, __dirname: __dirname }
-    const script  = new vm.Script(fs.readFileSync(__dirname + "/lib/api-stubs.js", 'utf8'))
-    var ctx = vm.createContext(glob);
+    var script  = new vm.Script(fs.readFileSync(__dirname + "/lib/timeline-model.js", 'utf8'))
+    var ctx = vm.createContext(glob)
     var output = script.runInContext(ctx)
 
     this.sandbox = ctx.instance;
@@ -59,4 +55,4 @@ class TraceToTimelineModel {
   }
 }
 
-module.exports = TraceToTimelineModel
+module.exports = ModelAPI
