+++ lib/api-stubs.js
@@ -1,21 +1,6 @@
 /* global WebInspector Runtime WorkerRuntime Protocol */
-"use strict"
-
-var fs = require('fs');
 
 var noop = function () {}
-
-this.global = this
-this.self = global
-this.window = global
-this.console = console
-this.WebInspector = {}; // might prefer being a var
-
-this.Runtime = {}
-this.TimelineModelTreeView = {};
-this.TreeElement = {}
-this.WorkerRuntime = {}
-this.Protocol = {}
 
 WebInspector.targetManager = {}
 WebInspector.targetManager.observeTargets = noop
@@ -36,123 +21,3 @@ Runtime.experiments = {}
 WorkerRuntime.Worker = noop
 WebInspector.TimelineTreeView = {}
 
-function req(path){
-  var filesrc = fs.readFileSync(__dirname + '/node_modules/' + path, 'utf8');
-  eval(filesrc)
-}
-
-req('chrome-devtools-frontend/front_end/common/Object.js')
-req('chrome-devtools-frontend/front_end/common/SegmentedRange.js')
-req('chrome-devtools-frontend/front_end/platform/utilities.js')
-req('chrome-devtools-frontend/front_end/sdk/Target.js')
-req('chrome-devtools-frontend/front_end/bindings/TempFile.js')
-req('chrome-devtools-frontend/front_end/sdk/TracingModel.js')
-req('chrome-devtools-frontend/front_end/timeline/TimelineJSProfile.js')
-req('chrome-devtools-frontend/front_end/timeline/TimelineUIUtils.js')
-req('chrome-devtools-frontend/front_end/sdk/CPUProfileDataModel.js')
-req('chrome-devtools-frontend/front_end/timeline/LayerTreeModel.js')
-req('chrome-devtools-frontend/front_end/timeline/TimelineModel.js')
-req('chrome-devtools-frontend/front_end/timeline/TimelineTreeView.js')
-req('chrome-devtools-frontend/front_end/ui_lazy/SortableDataGrid.js')
-req('chrome-devtools-frontend/front_end/timeline/TimelineProfileTree.js')
-req('chrome-devtools-frontend/front_end/components_lazy/FilmStripModel.js')
-req('chrome-devtools-frontend/front_end/timeline/TimelineIRModel.js')
-req('chrome-devtools-frontend/front_end/timeline/TimelineFrameModel.js')
-
-// minor configurations
-req('../lib/devtools-init.js')
-// polyfill the bottom-up and topdown tree sorting
-req('../lib/timeline-model-treeview.js')
-
-class SandboxedModel {
-
-  constructor() {
-    this.WI = WebInspector;
-  }
-  init(events) {
-    // (devtools) tracing model
-    this._tracingModel = new WebInspector.TracingModel(new WebInspector.TempFileBackingStorage('tracing'))
-    // timeline model
-    this._timelineModel = new WebInspector.TimelineModel(WebInspector.TimelineUIUtils.visibleEventsFilter())
-
-    // populate with events
-    this._tracingModel.reset()
-    this._tracingModel.addEvents(typeof events === 'string' ? JSON.parse(events) : events)
-    this._tracingModel.tracingComplete()
-    this._timelineModel.setEvents(this._tracingModel)
-
-    this._aggregator = new WebInspector.TimelineAggregator((event) => WebInspector.TimelineUIUtils.eventStyle(event).category.name)
-
-    return this;
-  }
-
-  timelineModel() {
-    return this._timelineModel;
-  }
-
-  tracingModel() {
-    return this._tracingModel;
-  }
-
-  topDown() {
-    var filters = [];
-    filters.push(WebInspector.TimelineUIUtils.visibleEventsFilter());
-    filters.push(new WebInspector.ExcludeTopLevelFilter());
-    var nonessentialEvents = [
-      WebInspector.TimelineModel.RecordType.EventDispatch,
-      WebInspector.TimelineModel.RecordType.FunctionCall,
-      WebInspector.TimelineModel.RecordType.TimerFire
-    ];
-    filters.push(new WebInspector.ExclusiveNameFilter(nonessentialEvents));
-
-    return WebInspector.TimelineProfileTree.buildTopDown(this._timelineModel.mainThreadEvents(),
-                  filters, /* startTime */ 0, /* endTime */ Infinity, WebInspector.TimelineAggregator.eventId)
-  }
-
-  bottomUp() {
-    var topDown = this.topDown();
-    var noGrouping = WebInspector.TimelineAggregator.GroupBy.None
-    var noGroupAggregator =  this._aggregator.groupFunction(noGrouping)
-    return WebInspector.TimelineProfileTree.buildBottomUp(topDown, noGroupAggregator)
-  }
-
-  // @ returns a grouped and sorted tree
-  bottomUpGroupBy(grouping) {
-    var topDown = this.topDown();
-    var groupSetting = WebInspector.TimelineAggregator.GroupBy[grouping] // one of: None Category Subdomain Domain URL
-    var groupURLAggregator =  this._aggregator.groupFunction(groupSetting)
-    var bottomUpGrouped = WebInspector.TimelineProfileTree.buildBottomUp(topDown, groupURLAggregator)
-    // sort the grouped tree, in-place
-    new TimelineModelTreeView(bottomUpGrouped).sortingChanged('self', 'desc')
-    return bottomUpGrouped
-  }
-
-  frameModel() {
-    var frameModel = new WebInspector.TracingTimelineFrameModel()
-    frameModel.addTraceEvents({ /* target */ }, this._timelineModel.inspectedTargetEvents(), this._timelineModel.sessionId() || '')
-    return frameModel
-  }
-
-  filmStripModel() {
-    return new WebInspector.FilmStripModel(this._tracingModel)
-  }
-
-
-  interactionModel() {
-    var irModel = new WebInspector.TimelineIRModel()
-    irModel.populate(this._timelineModel)
-    return irModel
-  }
-
-}
-
-// var filesrc = fs.readFileSync(__dirname + "/../test/devtools-homepage-w-screenshots-trace.json", 'utf8');
-
-var instance = new SandboxedModel();
-
-// instance.init(filesrc)
-
-// module.exports = instance
-
-
-// return new SandboxedModel();
