+++ lib/devtools-monkeypatches.js
